+++ ui-common/src/main/java/com/example/testapp/uicommon/design/PracticeExamTopBarMetrics.kt
@@ -3,6 +3,6 @@ package com.example.testapp.uicommon.design
 import androidx.compose.ui.unit.dp
 
 object PracticeExamTopBarMetrics {
-    val barHeight = 64.dp
-    val iconButtonSize = 40.dp
+    val barHeight = 72.dp
+    val iconButtonSize = 44.dp
 }